Updated list view of app and added the abilit to add drinks to those lists.

+++ MyTab/app/src/main/java/com/example/chandler/hack/IDrink.java
@@ -16,9 +16,9 @@ public interface IDrink {
 
 	public Image getImage();
 
-	public void setRating(int rating);
+	public void setRating(float rating);
 
-	public int getRating();
+	public float getRating();
 
 	public void update();
